+++ R4_STM_CODE/Core/Src/Igniter.cpp
@@ -15,7 +15,7 @@ bool Igniter::isConnected() const{
 		return 0;
 }
 
-void Igniter::FIRE(uint8_t milisecs){
+void Igniter::FIRE(uint16_t milisecs){
 	HAL_GPIO_WritePin(GPIO_PORT_IGNITER, PIN_IGNITER, GPIO_PIN_SET);
 	HAL_Delay(milisecs);
 	HAL_GPIO_WritePin(GPIO_PORT_IGNITER, PIN_IGNITER, GPIO_PIN_RESET);
